src/main/java/model/Client.java
package model;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

import java.util.UUID;

@Document(collection = "clients")
public class Client {

    @Id
    private UUID id;
    private String nom;
    private String email;

    public Client() {}

    public Client(String nom, String email) {
        this.id = UUID.randomUUID();
        this.nom = nom;
        this.email = email;
    }

    public UUID getId() {
        return id;
    }

    public void setId(UUID id) {
        this.id = id;
    }

    public String getNom() {
        return nom;
    }

    public void setNom(String nom) {
        this.nom = nom;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }
}
